[BE] refactor : rebase 중복 메서드 제거
