+++ backend/src/main/java/moim_today/implement/moim/joined_moim/JoinedMoimFinder.java
@@ -2,7 +2,6 @@ package moim_today.implement.moim.joined_moim;
 
 import moim_today.global.annotation.Implement;
 import moim_today.global.error.ForbiddenException;
-import moim_today.global.error.NotFoundException;
 import moim_today.persistence.entity.moim.joined_moim.JoinedMoimJpaEntity;
 import moim_today.persistence.repository.moim.joined_moim.JoinedMoimRepository;
 import org.springframework.transaction.annotation.Transactional;
@@ -10,8 +9,6 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 
 import static moim_today.global.constant.exception.MoimExceptionConstant.MOIM_FORBIDDEN_ERROR;
-
-import static moim_today.global.constant.exception.JoinedMoimExceptionConstant.JOINED_MOIM_MEMBER_NOT_FOUNT;
 
 @Implement
 public class JoinedMoimFinder {
@@ -35,15 +32,7 @@ public class JoinedMoimFinder {
     @Transactional(readOnly = true)
     public void validateMemberInMoim(final long moimId, final long memberId) {
         boolean isMemberInMoim = joinedMoimRepository.existsByMoimIdAndMemberId(moimId, memberId);
-        if(!isMemberInMoim){
-            throw new NotFoundException(JOINED_MOIM_MEMBER_NOT_FOUNT.message());
-        }
-    }
-
-    @Transactional(readOnly = true)
-    public void validateJoinedMember(final long memberId, final long moimId) {
-        List<Long> joinedMemberIds = joinedMoimRepository.findAllJoinedMemberId(moimId);
-        if (!joinedMemberIds.contains(memberId)) {
+        if (!isMemberInMoim) {
             throw new ForbiddenException(MOIM_FORBIDDEN_ERROR.message());
         }
     }
